Revert "Move PublishModal to its own directory"
This reverts commit 177f9b1643ed672c07de5b8f3f42c3b8513cadbe.
