+++ contentcuration/contentcuration/frontend/channelEdit/components/publish/PublishModal.vue
